Weather oracles completed.

+++ test/contracts/PullOracleConsumers/CentralizedPullOracleConsumer.sol
@@ -4,7 +4,7 @@ import "../../../contracts/Interfaces/PullOracle.sol";
 
 contract CentralizedPullOracleConsumer {
   PullOracle public oracle;
-  int8 public resolution;
+  int public resolution;
 
   constructor(PullOracle _oracle) public {
     oracle = _oracle;
@@ -12,6 +12,6 @@ contract CentralizedPullOracleConsumer {
 
   function getResult(bytes32 id) public {
     bytes32 result = oracle.resultFor(id);
-    resolution = int8(result);
+    resolution = int(result);
   }
 }
